+++ package.json
@@ -72,7 +72,6 @@
   },
   "scripts": {
     "build": "npm run build-js",
-    "build-css": "cat src/**/*.css src/*.css > dist/react-sticky-table.css",
     "build-js": "babel --plugins '@babel/plugin-transform-modules-umd' src --ignore __tests__ --out-dir ./dist",
     "build-storybook": "build-storybook -o ./docs",
     "prepare": "npm run build",
