+++ src/__tests__/Cell.js
@@ -4,7 +4,7 @@ import {mount, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import {expect} from 'chai';
 
-import {Cell} from '../../index';
+import {Cell} from '../index';
 
 configure({ adapter: new Adapter() });
 
@@ -17,6 +17,6 @@ describe('Cell', () => {
       </Cell>
     );
 
-    expect(cell.find('.sticky-table-cell span')).to.have.length(2);
+    expect(cell.find(Cell).find('span')).to.have.length(2);
   });
 });
