+++ src/__tests__/Table.js
@@ -4,7 +4,7 @@ import {mount, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 import {expect} from 'chai';
 
-import {Table, Row, Cell} from '../../index';
+import {Table, Row, Cell} from '../index';
 
 configure({ adapter: new Adapter() });
 
@@ -13,13 +13,13 @@ describe('Table', () => {
     const table = mount(
       <Table>
         <Row>
-          <Cell>{'header 1'}</Cell>
-          <Cell>{'header 2'}</Cell>
+          <Cell>header 1</Cell>
+          <Cell>header 2</Cell>
         </Row>
       </Table>
     );
 
-    expect(table.find('.sticky-table-row')).to.have.length(1);
-    expect(table.find('.sticky-table-cell')).to.have.length(2);
+    expect(table.find(Row)).to.have.length(1);
+    expect(table.find(Cell)).to.have.length(2);
   });
 });
